requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]

# Database and ORM
sqlalchemy
psycopg2-binary
supabase

# Security and Authentication (The "C" and "I" of CIA)
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv

# Frontend Dashboard
streamlit
requests
pandas